refactor: extract as useEmail hook #38
