+++ src/hooks/index.ts
@@ -7,3 +7,4 @@ export * from './useAtomInput';
 export * from './useCorrect';
 export * from './useCardMutation';
 export * from './useOutsideClick';
+export * from './useEmail';
